# openapi.json
{
  "components": {
    "schemas": {
      "GetRequest": {
        "properties": {
          "artifact_id": {
            "minLength": 1,
            "title": "Artifact Id",
            "type": "string"
          },
          "repo_fingerprint": {
            "minLength": 1,
            "title": "Repo Fingerprint",
            "type": "string"
          },
          "symbol_id": {
            "minLength": 1,
            "title": "Symbol Id",
            "type": "string"
          },
          "workspace_id": {
            "minLength": 1,
            "title": "Workspace Id",
            "type": "string"
          }
        },
        "required": [
          "workspace_id",
          "repo_fingerprint",
          "artifact_id",
          "symbol_id"
        ],
        "title": "GetRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "IndexRequest": {
        "properties": {
          "repo_fingerprint": {
            "minLength": 1,
            "title": "Repo Fingerprint",
            "type": "string"
          },
          "repo_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Repo Ref"
          },
          "workspace_id": {
            "minLength": 1,
            "title": "Workspace Id",
            "type": "string"
          }
        },
        "required": [
          "workspace_id",
          "repo_fingerprint"
        ],
        "title": "IndexRequest",
        "type": "object"
      },
      "OutlineRequest": {
        "properties": {
          "artifact_id": {
            "minLength": 1,
            "title": "Artifact Id",
            "type": "string"
          },
          "repo_fingerprint": {
            "minLength": 1,
            "title": "Repo Fingerprint",
            "type": "string"
          },
          "workspace_id": {
            "minLength": 1,
            "title": "Workspace Id",
            "type": "string"
          }
        },
        "required": [
          "workspace_id",
          "repo_fingerprint",
          "artifact_id"
        ],
        "title": "OutlineRequest",
        "type": "object"
      },
      "QueryRequest": {
        "properties": {
          "artifact_id": {
            "minLength": 1,
            "title": "Artifact Id",
            "type": "string"
          },
          "mode": {
            "default": "search",
            "enum": [
              "search",
              "get",
              "outline"
            ],
            "title": "Mode",
            "type": "string"
          },
          "query": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Query"
          },
          "repo_fingerprint": {
            "minLength": 1,
            "title": "Repo Fingerprint",
            "type": "string"
          },
          "symbol_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Symbol Id"
          },
          "workspace_id": {
            "minLength": 1,
            "title": "Workspace Id",
            "type": "string"
          }
        },
        "required": [
          "workspace_id",
          "repo_fingerprint",
          "artifact_id"
        ],
        "title": "QueryRequest",
        "type": "object"
      },
      "SearchRequest": {
        "properties": {
          "artifact_id": {
            "minLength": 1,
            "title": "Artifact Id",
            "type": "string"
          },
          "query": {
            "minLength": 1,
            "title": "Query",
            "type": "string"
          },
          "repo_fingerprint": {
            "minLength": 1,
            "title": "Repo Fingerprint",
            "type": "string"
          },
          "workspace_id": {
            "minLength": 1,
            "title": "Workspace Id",
            "type": "string"
          }
        },
        "required": [
          "workspace_id",
          "repo_fingerprint",
          "artifact_id",
          "query"
        ],
        "title": "SearchRequest",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "voro-index-guard",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/health": {
      "get": {
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health"
      }
    },
    "/v1/get": {
      "post": {
        "operationId": "get_symbol_v1_get_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Symbol"
      }
    },
    "/v1/index": {
      "post": {
        "operationId": "create_index_v1_index_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IndexRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Index"
      }
    },
    "/v1/outline": {
      "post": {
        "operationId": "get_outline_v1_outline_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OutlineRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Outline"
      }
    },
    "/v1/query": {
      "post": {
        "operationId": "query_index_v1_query_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Query Index"
      }
    },
    "/v1/search": {
      "post": {
        "operationId": "search_index_v1_search_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Index"
      }
    }
  }
}